añadiendo contentdialog para confirmar eliminar un producto

+++ PriceTrendCam/ViewModels/ProductDetailsViewModel.cs
@@ -2,6 +2,8 @@
 
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
 using PriceTrendCam.Contracts.ViewModels;
 using PriceTrendCam.Core.Contracts.Services;
 using PriceTrendCam.Core.Models;
@@ -60,21 +62,35 @@ public partial class ProductDetailsViewModel : ObservableRecipient, INavigationA
         }
     }
     [RelayCommand]
-    private async void DeleteProduct()
+    private async void DeleteProduct(XamlRoot root)
     {
-        //TODO: CONFIRMAR ELIMINAR LA STORE PARA QUE NO SE BORRE INSTANTANEAMENTE
-        await App.PriceTrackerService.DeleteAsync(Selected);
-
-        SampleItems.Clear();
-
-        // TODO: Replace with real data.
-        var data = await _sampleDataService.GetListDetailsDataAsync();
-
-        foreach (var item in data)
+        ContentDialog dialog = new()
         {
-            SampleItems.Add(item);
-        }
+            Title = "Delete Store",
+            XamlRoot = root,
+            CloseButtonText = "Close",
+            DefaultButton = ContentDialogButton.Close,
+            PrimaryButtonText = "Delete",
+            SecondaryButtonText = "Cancel",
+            Content = "Are you sure you want to delete the store?"
+        };
 
-        EnsureItemSelected();
+        await dialog.ShowAsync().AsTask();
+
+        if (dialog.IsPrimaryButtonEnabled)
+        {
+            await App.PriceTrackerService.DeleteAsync(Selected);
+
+            SampleItems.Clear();
+
+            // TODO: Replace with real data.
+            var data = await _sampleDataService.GetListDetailsDataAsync();
+
+            foreach (var item in data)
+            {
+                SampleItems.Add(item);
+            }
+            EnsureItemSelected();
+        }
     }
 }
